pyproject.toml
[project]
name = "mlx-yap"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "mlx-whisper>=0.4",
    "sounddevice>=0.5",
    "numpy>=2.0",
    "pynput>=1.7",
    "mistralai>=1.0,<2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.scripts]
yap = "mlx_yap.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mlx_yap"]

[tool.ruff]
line-length = 100
target-version = "py312"
